foo-4: review fix v2
