+++ src/main/resources/application.yaml
@@ -1,24 +1,21 @@
 database:
   name: quisell
-  username:
-  password:
-  host:
+  username: postgres
+  password: postgres
+  host: localhost
   port: 5432
-
-psql:
-  user: postgres
-  password: pass
-  schema: public
 
 spring:
   jpa:
     hibernate:
       ddl-auto: none
   liquibase:
-    change-log: classpath:db/changelog/liquibase-changelog.yml
+    change-log: classpath:db/changelog-master.yml
     enabled: true
     url: ${spring.datasource.url}
+    user: ${database.username}
+    password: ${database.password}
   datasource:
-    url: jdbc:postgresql://${database.host}:${database.port}/${database.name}?currentSchema=${psql.schema}&user=${psql.user}&password=${psql.password}
-    username:
-    password:
+    url: jdbc:postgresql://${database.host}:${database.port}/${database.name}
+    username: ${database.username}
+    password: ${database.password}
