+++ src/main/resources/db/changelog/changelog-0.0.1.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+         http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.1.xsd">
+
+    <changeSet id="01" author="ognevoydev">
+
+        <createTable tableName="posts">
+            <column name="id" type="uuid">
+                <constraints nullable="false" primaryKey="true"/>
+            </column>
+            <column name="title" type="varchar(255)">
+                <constraints nullable="false"/>
+            </column>
+            <column name="description" type="text">
+                <constraints nullable="false"/>
+            </column>
+            <column name="price" type="int">
+                <constraints/>
+            </column>
+            <column name="used" type="boolean">
+                <constraints nullable="false"/>
+            </column>
+            <column name="phone" type="char(10)">
+                <constraints nullable="false"/>
+            </column>
+            <column name="created_at" type="timestamp">
+                <constraints nullable="false"/>
+            </column>
+            <column name="updated_at" type="timestamp">
+                <constraints/>
+            </column>
+            <column name="deleted_at" type="timestamp">
+                <constraints/>
+            </column>
+        </createTable>
+
+    </changeSet>
+</databaseChangeLog>
