--- src/main/resources/db/changelog/create-posts-schema.yml
@@ -1,48 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: 01
-      author: ognevoy
-      changes:
-        - createTable:
-            tableName: posts
-            columns:
-              - column:
-                  name: id
-                  type: uuid
-                  constraints:
-                    primaryKey: true
-                    nullable: false
-              - column:
-                  name: title
-                  type: varchar(255)
-                  constraints:
-                    nullable: false
-              - column:
-                  name: description
-                  type: text
-                  constraints:
-                    nullable: false
-              - column:
-                  name: price
-                  type: int
-              - column:
-                  name: used
-                  type: boolean
-                  constraints:
-                    nullable: false
-              - column:
-                  name: phone
-                  type: char(10)
-                  constraints:
-                    nullable: false
-              - column:
-                  name: created_at
-                  type: timestamp
-                  constraints:
-                    nullable: false
-              - column:
-                  name: updated_at
-                  type: timestamp
-              - column:
-                  name: deleted_at
-                  type: timestamp
