+++ src/main/resources/db/changelog/liquibase-changelog.yml
@@ -1,3 +1,3 @@
 databaseChangeLog:
   - include:
-      file: db/changelog/create-posts-schema.yml
+      file: db/changelog/changelog-0.0.1.xml
